+++ src/export_runtime/__init__.py
@@ -16,7 +16,25 @@ from .payloads import ExportPayload, ExportPayloadLine, ExportPayloadParty, payl
 from .ports import ExportAdapterPort
 from .policy import ExportIdempotencyPolicy, payload_invalid_readiness_issue, payload_readiness_issues
 from .readiness import ExportReadinessIssue, ExportReadinessResult, readiness_result
+from .queries import (
+    DEFAULT_EXPORT_QUERY_LIMIT,
+    MAX_EXPORT_QUERY_LIMIT,
+    MAX_EXPORT_QUERY_OFFSET,
+    ExportAttemptQuery,
+    ExportPage,
+    ExportPageRequest,
+)
+from .repositories import (
+    ExportAttemptReadRepository,
+    ExportAttemptWriteRepository,
+    ExportRepositoryReader,
+    ExportRepositoryWriter,
+    ExportResultReadRepository,
+    ExportResultWriteRepository,
+)
+from .repository_errors import ExportRepositoryError, ExportRepositoryErrorCode
 from .results import ExportAdapterResult, ExportResult
+from .store import ACTIVE_EXPORT_STATUSES, TERMINAL_EXPORT_STATUSES, InMemoryExportStore
 from .statuses import (
     EXPORT_OPERATION_STATUS_VALUES,
     EXPORT_READINESS_ISSUE_CODES,
@@ -32,10 +50,17 @@ __all__ = [
     "EXPORT_OPERATION_STATUS_VALUES",
     "EXPORT_READINESS_ISSUE_CODES",
     "EXPORT_STATUS_VALUES",
+    "ACTIVE_EXPORT_STATUSES",
+    "DEFAULT_EXPORT_QUERY_LIMIT",
+    "MAX_EXPORT_QUERY_LIMIT",
+    "MAX_EXPORT_QUERY_OFFSET",
     "PAYLOAD_FINGERPRINT_DOMAIN",
     "ExportAdapterPort",
     "ExportAdapterResult",
     "ExportAttempt",
+    "ExportAttemptQuery",
+    "ExportAttemptReadRepository",
+    "ExportAttemptWriteRepository",
     "ExportAuditIntent",
     "ExportError",
     "ExportErrorCode",
@@ -51,14 +76,24 @@ __all__ = [
     "ExportPayloadLine",
     "ExportPayloadParty",
     "ExportPermission",
+    "ExportPage",
+    "ExportPageRequest",
     "ExportIdempotencyPolicy",
     "ExportReadinessIssue",
     "ExportReadinessIssueCode",
     "ExportReadinessResult",
     "ExportResult",
+    "ExportRepositoryError",
+    "ExportRepositoryErrorCode",
+    "ExportRepositoryReader",
+    "ExportRepositoryWriter",
+    "ExportResultReadRepository",
+    "ExportResultWriteRepository",
     "ExportStatus",
     "ExportTarget",
     "ExportTargetType",
+    "InMemoryExportStore",
+    "TERMINAL_EXPORT_STATUSES",
     "build_export_payload",
     "canonical_payload_json",
     "fingerprint_export_payload",